Reduce the flakiness of the test case to something more reasonable.
The original chance of occurrence of a failure of the test case of
random number generation was too high. Thus I changed the thresholds
considered to evaluate the number.

Another improvement was in the random number generator. The previous
code would never generate a number in the upper limit. This has been
fixed. The 'get_random_integer_number' was copied from StackOverflow
(https://stackoverflow.com/a/29246176).